% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_DescribeRecord}
\alias{CSW_DescribeRecord}
\title{Get Description of CSW Records}
\usage{
CSW_DescribeRecord(typeName = c("csw:Record", "gmd:MD_Metadata",
  "gmi:MI_Metadata"), typeNames = c("csw:Record", "gmd:MD_Metadata",
  "gmi:MI_Metadata"),
  namespace = c("csw:http://www.opengis.net/cat/csw/2.0.2",
  "gmd:http://www.isotc211.org/2005/gmd",
  "gmi:http://www.isotc211.org/2005/gmi"),
  outputFormat = "application/xml", version = CSW_get_version(),
  baseurl = CSW_get_url(), verbose = c("N", "F", "Y"))
}
\arguments{
\item{typeName}{Character vector with one or both of \code{csw:Record} and \code{gmd:MD_Metadata}. But see remarks.}

\item{typeNames}{alternative for \code{typeName}.}

\item{namespace}{Character vector with one or both of \code{csw:http://www.opengis.net/cat/csw/2.0.2} and \code{gmd:http://www.isotc211.org/2005/gmd}. See Remarks for the relation between \code{typeName} and \code{namespace}.}

\item{outputFormat}{Character vector. Only  \code{application/xml} is apparently (?) supported by CSW}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
An \code{xml} document with the record description in the indicated layout.
}
\description{
This function gives a description of the catalog records in the form of an \code{xml} document. With typeName \code{csw:Record}  the description follows a table format and that is why I like this format more than the one you get with typeName \code{gmd:MD_Metadata}. It is possible to get both formats in one \code{xml} document as shown in example exp12.
}
\section{Remarks}{

typeName \code{gfc:FC_FeatureCatalogue} and \code{dcat} were also mentioned as possible values but can not be combined with the two namespace options that are given here. Also found was the gmi analog for gmd but for the default catalog no results were delivered. Moreover the typeName and namespace must match: the first typeName with the first namespace and the same for the second one.
}

\examples{
\dontrun{
exp1  = CSW_DescribeRecord()
exp2  = CSW_DescribeRecord(typeName = 'gmd:MD_Metadata',
       namespace = 'gmd:http://www.isotc211.org/2005/gmd')
exp12 = CSW_DescribeRecord(
    typeName =
        c('csw:Record',	'gmd:MD_Metadata') ,
    namespace = c(
        'csw:http://www.opengis.net/cat/csw/2.0.2',
        'gmd:http://www.isotc211.org/2005/gmd') )
}
}
