% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CSW_display_node}
\alias{CSW_display_node}
\title{Display an xml document or node}
\usage{
CSW_display_node(n1, file = NULL)
}
\arguments{
\item{n1}{An object of type \code{xml_document} or \code{xml_node}}

\item{file}{Path to file or connection to write to. Default NULL i.e. write to console log}
}
\description{
This function displays the contents of an xml document or node in the console log.
}
\examples{
\dontrun{
CSW_display_node(xml1)
}
}
