% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseurl.R
\name{CSW_get_url}
\alias{CSW_get_url}
\title{Gets the url for the catalog structure}
\usage{
CSW_get_url()
}
\value{
Character vector with the active url.
}
\description{
Determines from which structure catalog information will be extracted. If the function CSW_set_url
is not called yet this will be done first with the default \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} (the url for the PDOK/NGR structures).#'
}
\examples{
CSW_get_url()
}
\seealso{
\code{\link{CSW_set_url}}
}
