% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseurl.R
\name{CSW_get_version}
\alias{CSW_get_version}
\title{Gets the version for the catalog structure}
\usage{
CSW_get_version()
}
\value{
Character vector with the active version.
}
\description{
Determines the version of the  structure catalog information that will be extracted. If the function CSW_set_version
is not called yet this will be done first with the default \code{2.0.2} (the version for the PDOK/NGR structures).
}
\examples{
CSW_get_version()
}
\seealso{
\code{\link{CSW_set_version}}
}
