% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetCapabilities}
\alias{CSW_GetCapabilities}
\title{Get Capabilities of the CSW server}
\usage{
CSW_GetCapabilities(sections = c("OperationsMetadata",
  "ServiceIdentification", "ServiceProvider", "Filter_Capabilities"),
  typeNames = c("csw:Record", "gmd:MD_Metadata",
  "gfc:FC_FeatureCatalogue", "dcat"), version = CSW_get_version(),
  baseurl = CSW_get_url(), verbose = c("N", "F", "Y"))
}
\arguments{
\item{sections}{Character vector to restrict the output to one of the \code{OperationsMetadata}, \code{ServiceIdentification}, \code{ServiceProvider} and \code{Filter_Capabilities} sections. When the argument is not specified all information is provided. NB whatever section is requested, the \code{Filter_Capabilities} section is always included.}

\item{typeNames}{Character vector with one of \code{csw:Record}, \code{gmd:MD_Metadata},     \code{gfc:FC_FeatureCatalogue} and \code{dcat}. Default \code{csw:Record}.}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
An \code{xml} document with the indicated server capabilities.
}
\description{
The \code{CSW_GetCapabilities} function gives information about the capabilities that are available to query the catalog. By calling this function without arguments we receive an \code{xml} document with this information. We can view the contents of this document (e.g.  with the utility function \code{CSW_display_node}) and study its structure. We see among other things the operations that are available and the parameters with which these operations can be called. Apart of viewing the document we can also select parts of it with functions of the \pkg{xml2} package by using \code{XPATH} expressions. The vignette shows how one can see which operations are provided by the CSW interface and which parameters are available for the GetRecords operation (i.e. the \code{CSW_GetRecords} function).
}
\examples{
\dontrun{
exp1  = CSW_GetCapabilities(sections = 'OperationsMetadata',
         typeNames = 'csw:Record')
# CSW_display_node(exp1)
exp2  = CSW_GetCapabilities(sections = 'OperationsMetadata',
         typeNames = 'gmd:MD_Metadata')
}
}
