% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetData}
\alias{CSW_GetData}
\title{Retrieve data from CSW server}
\usage{
CSW_GetData(request = c("GetCapabilities", "GetRecords",
  "DescribeRecord", "GetDomain", "GetRecordById", "Transaction",
  "Harvest"), version = CSW_get_version(), baseurl = CSW_get_url(),
  verbose = c("N", "F", "Y"), ...)
}
\arguments{
\item{request}{Character vector indicating the CSW operation that has to be performed. One of \code{GetCapabilities}, \code{GetRecords}, \code{DescribeRecord}, \code{GetDomain}, \code{GetRecordById}, \code{Transaction}, \code{Harvest}.  Default \code{GetCapabilities}.}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}

\item{...}{Optionally one or more \code{name=value} arguments. See example exp2.}
}
\value{
An \code{xml} document with the requested information.
}
\description{
This function is the only function of the CSW package that actually retrieves data from the CSW Server. Most often this function will not be used directly by the user. For the various CSW operations dedicated functions are provided that generate a string with parameters that is passed to \code{CSW_GetData}. With the parameter \code{verbose} one can indicate that this generated string is printed in the console log. These parameters can then specified for this function (e.g. when an option would be needed that is not provided for by the dedicated functions). See example.
}
\section{Remarks}{

\code{CSW_GetData} always returns an xml document where other functions sometimes try to reformat the information when requested. Using this function can help with debugging when there is a problem (e.g. with this formatting).
}

\examples{
\dontrun{
# exp1 and exp2 request the same information from the server:
exp1 = CSW_DescribeRecord(typeName = 'gmd:MD_Metadata',
	namespace = 'gmd:http://www.isotc211.org/2005/gmd',verbose='N')
exp2 = CSW_GetData(request='DescribeRecord',
	namespace ='gmd:http://www.isotc211.org/2005/gmd',
	typeName ='gmd:MD_Metadata',verbose='N')
}
}
