% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetDomain}
\alias{CSW_GetDomain}
\title{Get Domain values of the CSW server}
\usage{
CSW_GetDomain(ParameterName = "", PropertyName = "",
  output = c("list", "xml"), version = CSW_get_version(),
  baseurl = CSW_get_url(), verbose = c("N", "F", "Y"))
}
\arguments{
\item{ParameterName}{Character vector with one or more CSW parameters. See  \code{\link{CSW_GetDomainParameterNames}} for possible values. When \code{ParameterName} and \code{PropertyName} are both used in the same call \code{ParameterName} will be ignored.}

\item{PropertyName}{Character vector with one or more CSW parameters. See  \code{\link{CSW_GetQueryables}} for possible values. When \code{ParameterName} and \code{PropertyName} are both used in the same call \code{ParameterName} will be ignored.}

\item{output}{Character vector with one of \code{list} and \code{xml}. With the default \code{list} the \code{xml} output is converted to a list.}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
An xml_document or list with the values permitted (for \code{ParameterName}) or to be used (for \code{PropertyName})
}
\description{
With the \code{GetDomain} operation the CSW server can return information about \code{ParameterName}s or \code{PropertyName}s (fields). \cr\cr
We can inquire after the parameters of an operation by specifying the \code{ParameterName} argument. This argument consists of one or more \code{operation.parameter} pairs as in example \code{exp1}. The output is a list (or alternatively an xml_document) with the values that can be used for the parameters. In each pair the part before the point should be an operation; the part after the point a parameter.  \code{\link{CSW_GetDomainParameterNames}} gives the same information for all possible combinations (at the cost of some extra run-time). \cr\cr
We can also use the \code{CSW_GetDomain} function to inquire which values a certain field can take in a catalog. We do this by specifying the fieldname in the \code{PropertyName} argument. This argument consists of one or more fieldnames as e.g. in example \code{exp2}. The output is (just as in the \code{ParameterName} case) a list (or alternatively an xml_document) with the values that can be taken by the field. \code{\link{CSW_GetQueryables}} will show which fieldnames (\code{PropertyNames}) are recognized.
}
\examples{
\dontrun{
exp1 = CSW_GetDomain(
   ParameterName = "GetRecords.outputFormat,GetRecords.ElementSetName",
   verbose = "Y")
exp2 = CSW_GetDomain(
   PropertyName = "Language,GeographicDescriptionCode",
   verbose = "Y")
}
}
\seealso{
\code{\link{CSW_GetDomainParameterNames}}, \code{\link{CSW_GetQueryables}}
}
