% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetDomainParameterNames}
\alias{CSW_GetDomainParameterNames}
\title{Get elements that can be used in GetDomain as ParameterName}
\usage{
CSW_GetDomainParameterNames(version = CSW_get_version(),
  baseurl = CSW_get_url(), verbose = c("N", "F", "Y"))
}
\arguments{
\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
A \code{data.frame} with columns \code{n} (element names) and \code{v} (possible parameter values). An element name has the format CSW_operation.CSW_parameter.
}
\description{
\code{CSW_GetDomainParameterNames} calls CSW_GetDomain for all (im)possible combinations of
Operation and Parameter names assembles the result in a \code{data.frame}. Not all parameters are found in this way: e.g. \code{DescribeRecord.outputSchema} can be found in the output of the \code{GetCapabilities} output but not in the output of \code{CSW_GetDomainParameterNames} !?
}
\section{Remarks}{

I have seen examples where CSW_GetCapabilities directly shows
which combinations are supported by CSW_GetDomain
(see \url{https://docs.oracle.com/cd/E11882_01/appdev.112/e11830/sdo_csw.htm#SPATL963})
but for the default baseurl this is apparently not available.
}

\examples{
\dontrun{
exp1 = CSW_GetDomainParameterNames()
}
}
