% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetHits}
\alias{CSW_GetHits}
\title{Get the number of records that match a constraint}
\usage{
CSW_GetHits(constraint = "", constraintLanguage = "CQL_TEXT",
  constraint_language_version = "1.1.0", version = CSW_get_version(),
  baseurl = CSW_get_url(), verbose = c("N", "F", "Y"))
}
\arguments{
\item{constraint}{Character vector with a constraint written with `` and \code{constraint_language_version==1.1.0}.}

\item{constraintLanguage}{Character vector indicating in which language the constraint is written. Default \code{CQL_TEXT}.}

\item{constraint_language_version}{Character vector indicating the version of the constraint language. Default \code{1.1.0}.}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
An integer with the number of records that satisfy the constraint or an xml_document with the exception in case of an error
}
\description{
CSW_GetHits returns the number of records that satisfy a constraint. The default constraint language and its version are \code{CQL_TEXT} and \code{1.1.0} but can be overwritten by arguments of this function.
}
\examples{
\dontrun{
# total number of records in collection :
  exp1 = CSW_GetHits()
# total number of records related to 'duin*' (dune*)
  exp2 = CSW_GetHits(constraint="AnyText LIKE 'duin\%'" )
}
}
