% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetQueryables}
\alias{CSW_GetQueryables}
\title{Get Queryable values of the CSW server}
\usage{
CSW_GetQueryables(version = CSW_get_version(), baseurl = CSW_get_url(),
  verbose = c("N", "F", "Y"))
}
\arguments{
\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
A list with two sublists: one with the 'SupportedISOQueryables' and one with the 'AdditionalQueryables'
}
\description{
The function \code{CSW_GetQueryables} retrieves the names of properties (fields) that can be used in the \code{CSW_GetDomain} function or in a query. The output is a list with two sublists: one with the 'SupportedISOQueryables' and one with the 'AdditionalQueryables'.
}
\examples{
\dontrun{
exp1 = CSW_GetQueryables()
}
}
