% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetRecordById}
\alias{CSW_GetRecordById}
\title{Get catalog info for entry with given id}
\usage{
CSW_GetRecordById(id = "", typeNames = c("csw:Record",
  "gmd:MD_Metadata", "gfc:FC_FeatureCatalogue", "dcat"),
  outputSchema = c("http://www.opengis.net/cat/csw/2.0.2",
  "http://www.isotc211.org/2005/gmd", "http://www.isotc211.org/2005/gfc",
  "http://www.w3.org/ns/dcat#"), ElementSetName = c("summary", "full",
  "brief"), namespace = c("csw:http://www.opengis.net/cat/csw/2.0.2",
  "gmd:http://www.isotc211.org/2005/gmd"), output = c("table", "list",
  "xml"), version = CSW_get_version(), baseurl = CSW_get_url(),
  verbose = c("N", "F", "Y"))
}
\arguments{
\item{id}{Character vector with the Identifier of the catalog entry that has to be retrieved.}

\item{typeNames}{Character vector with one (or both?) of \code{csw:Record} and \code{gmd:MD_Metadata}. Also \code{gfc:FC_FeatureCatalogue} and \code{dcat} can be used.}

\item{outputSchema}{Character vector with one (or more?) of \code{http://www.opengis.net/cat/csw/2.0.2} , \code{http://www.isotc211.org/2005/gmd}, \code{http://www.isotc211.org/2005/gfc} and \code{http://www.w3.org/ns/dcat#}. The first two work and the first of them give a more table like presentation.}

\item{ElementSetName}{Character vector with one of \code{summary}, \code{full} and \code{brief} indicating how much data is presented. Default \code{summary}.}

\item{namespace}{Character vector with one or both of \code{csw:http://www.opengis.net/cat/csw/2.0.2} and \code{gmd:http://www.isotc211.org/2005/gmd}. See Remarks for the relation between \code{typeName} and \code{namespace}.}

\item{output}{Character vector with one of \code{table}, \code{list} and \code{xml}. With \code{table} (the default) the results will be put in a (one row) data.frame. With \code{list} the output is conforming to that of  \code{\link{CSW_GetRecords}} when \code{list} is requested (the number of retrieved (here 1) and total records (here 1) are gathered in a list). With \code{xml} the retrieved xml_document is passed through without conversion.}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
An \code{xml} document with the record description in the indicated layout.
}
\description{
The only difference with \code{\link{CSW_GetRecords}} is that the \code{constraint} and a 'results\code{arguments of \code{\link{CSW_GetRecords}} are replace by the}id` argument of \code{\link{CSW_GetRecordById}}. Therefore see \code{\link{CSW_GetRecords}} for details.
}
\section{Remarks}{

We have tried various combinations of typeNames (first two only), outputSchema (same) and namespace.  These have only two possible outcomes and depend only on outputSchema: the csw output is more table like and the gmd output is more verbose (?)
}

\examples{
\dontrun{
 exp1  = CSW_GetRecordById(
          id = "ecb96a41-19da-4a57-b570-05a1ee5743ca",
          ElementSetName = "brief",
          output = "xml")
}
}
\seealso{
\code{\link{CSW_GetRecords}}
}
