% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetRecords}
\alias{CSW_GetRecords}
\title{Get catalog info for records satisfying constraint}
\usage{
CSW_GetRecords(constraint = "", constraintLanguage = "CQL_TEXT",
  constraint_language_version = "1.1.0", resultType = c("results",
  "validate"), typeNames = c("csw:Record", "gmd:MD_Metadata",
  "gfc:FC_FeatureCatalogue", "dcat"),
  outputSchema = c("http://www.opengis.net/cat/csw/2.0.2",
  "http://www.isotc211.org/2005/gmd", "http://www.isotc211.org/2005/gfc",
  "http://www.w3.org/ns/dcat#"), ElementSetName = c("summary", "full",
  "brief"), namespace = c("csw:http://www.opengis.net/cat/csw/2.0.2",
  "gmd:http://www.isotc211.org/2005/gmd"), maxRecords = "",
  startPosition = "", output = c("table", "list", "xml"),
  version = CSW_get_version(), baseurl = CSW_get_url(),
  verbose = c("N", "F", "Y"))
}
\arguments{
\item{constraint}{Character vector with a constraint written with \code{constraintLanguage=CQL_TEXT} and \code{constraint_language_version==1.1.0}.}

\item{constraintLanguage}{Character vector indicating in which language the constraint is written. Default \code{CQL_TEXT}.}

\item{constraint_language_version}{Character vector indicating the version of the constraint language. Default \code{1.1.0}.}

\item{resultType}{Character vector with one of \code{results} and \code{validate}. With \code{results} actual search results will be returned. With \code{validate} a Boolean will be returned indicating if the constraint is valid.}

\item{typeNames}{Character vector with one (or both?) of \code{csw:Record} and \code{gmd:MD_Metadata}. Also \code{gfc:FC_FeatureCatalogue} and \code{dcat} can be used.}

\item{outputSchema}{Character vector with one (or more?) of \code{http://www.opengis.net/cat/csw/2.0.2} , \code{http://www.isotc211.org/2005/gmd}, \code{http://www.isotc211.org/2005/gfc} and \code{http://www.w3.org/ns/dcat#}. The first two work and the first of them give a more table like presentation.}

\item{ElementSetName}{Character vector with one of \code{summary}, \code{full} and \code{brief} indicating how much data is presented. Default \code{summary}.}

\item{namespace}{Character vector with one or both of \code{csw:http://www.opengis.net/cat/csw/2.0.2} and \code{gmd:http://www.isotc211.org/2005/gmd}. See Remarks for the relation between \code{typeName} and \code{namespace}.}

\item{maxRecords}{Integer indicating the number of records that will be read from the server. Default is determined by the CSW server (for the default \code{baseurl} the default \code{maxRecords} is 10 but values in the order of 100 are accepted )}

\item{startPosition}{Integer indicating the number of the first record that will be returned. The default is 1. To read e.g. the records of a constraint with 150 hits one can issue two requests: the first one with \code{maxRecords = 100} and \code{startPosition = 1} (returning 100 records) and the second with  \code{maxRecords = 100} and \code{startPosition = 101} (that will return the last 50 records)}

\item{output}{Character vector with one of \code{table}, \code{list} and \code{xml}. With \code{table} (the default) the results will be put in a data.frame. With \code{list} this table is placed in a list with the number of retrieved and total number of records satisfying the constraint. With \code{xml} the retrieved xml_document is passed through without conversion.}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}
}
\value{
An \code{xml} document with the record description in the indicated layout.
}
\description{
The \code{CSW_GetRecords} function retrieves catalog entries from the CSW server (unless \code{resultType='validate'} is specified). The default output format is a \code{data.frame} with as alternatives an \code{xml_document} or \code{list} (see \code{output}). The fields of the \code{data.frame} are determined by the \code{ElementSetName} argument that results in few (\code{brief}), more (\code{summary}) and most (\code{full}) fields. The relation between \code{typeNames}, \code{outputSchema} and \code{namespace} is not clear to me, but it seems that the argument values must correspond: taking the first option for \code{typeNames} one should also take the first one for \code{outputSchema} and \code{namespace}. Also see remarks.\cr
The default constraint language and its version are \code{CQL_TEXT} and \code{1.1.0} but can be overwritten by arguments of this function. See the vignette for examples and references for the query language.
}
\section{Remarks}{

We have tried various combinations of typeNames (first two only), outputSchema (same) and namespace.  These have only two possible outcomes and depend only on outputSchema: the csw output is more table like and the gmd output is more verbose (?)
}

\examples{
\dontrun{
 exp1 = CSW_GetRecords(constraint="AnyText LIKE 'duin\%'" )
}
}
