% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{CSW_GetRecordsAll}
\alias{CSW_GetRecordsAll}
\title{Get catalog info for all records satisfying constraint}
\usage{
CSW_GetRecordsAll(constraint, typeNames = c("csw:Record",
  "gfc:FC_FeatureCatalogue", "dcat", "gmd:MD_Metadata"),
  ElementSetName = c("summary", "full", "brief"), output = c("table",
  "list"), batchsize = 100, version = "2.0.2", verbose = F,
  baseurl = CSW_get_url())
}
\arguments{
\item{constraint}{Character vector with a constraint written with \code{constraintLanguage=CQL_TEXT} and \code{constraint_language_version==1.1.0}.}

\item{typeNames}{Character vector with one (or both?) of \code{csw:Record} and \code{gmd:MD_Metadata}. Also \code{gfc:FC_FeatureCatalogue} and \code{dcat} can be used.}

\item{ElementSetName}{Character vector with one of \code{summary}, \code{full} and \code{brief} indicating how much data is presented. Default \code{summary}.}

\item{output}{Character vector with one of \code{table} and \code{list}. With \code{table} (the default) the results will be put in a data.frame. With \code{list} this table is placed in a list with the number of retrieved and total number of records satisfying the constraint.}

\item{batchsize}{Integer with the number of records that will be retrieved in each iteration of the retrieval loop. Default 100.}

\item{version}{Character vector with CSW version. Default is \code{2.0.2} but this can be changed for the remainder of the session with \code{\link{CSW_set_version}} .}

\item{verbose}{Character \code{N}, \code{F} or \code{Y}. When \code{F} the full generated and encoded url will be displayed, when \code{Y} the variable part of the url (without baseurl, service indication \code{CSW} and version) will be displayed in decoded form and when \code{N} (default) nothing of the generated url will be displayed.}

\item{baseurl}{Character vector with base url of the CSW server. Default is \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} but this can be changed for the remainder of the session with  \code{\link{CSW_set_url}} .}
}
\value{
An \code{xml} document with the record description in the indicated layout.
}
\description{
CSW_GetRecordsAll retrieves all records that satisfy a constraint by looping over CSW_GetRecords until all records are handled
}
\examples{
\dontrun{
 exp1 = CSW_GetRecordsAll(constraint="AnyText LIKE 'duin\%'" )
}
}
\seealso{
\code{\link{CSW_GetRecords}}
}
