% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseurl.R
\name{CSW_set_url}
\alias{CSW_set_url}
\title{Sets the url for CSW catalog structure}
\usage{
CSW_set_url(url = NULL)
}
\arguments{
\item{url}{NULL for the default or the url of the structure otherwise}
}
\value{
Character vector with (invisible) the url.
}
\description{
Determines from which structure catalog information  will be extracted. If this function is not called
the default \url{http://nationaalgeoregister.nl/geonetwork/srv/dut/csw?} (the url for the PDOK/NGR structures) will be used
}
\examples{
CSW_set_url("http://nationaalgeoregister.nl/geonetwork/srv/dut/inspire?")
}
\seealso{
\code{\link{CSW_get_url}}
}
