% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseurl.R
\name{CSW_set_version}
\alias{CSW_set_version}
\title{Sets the version for CSW catalog structure}
\usage{
CSW_set_version(version = NULL)
}
\arguments{
\item{version}{NULL for the default or the version of the structure otherwise}
}
\value{
Character vector with (invisible) the active version.
}
\description{
Determines the version of the  structure catalog information that will be extracted. If this function is not called
the default \code{2.0.2} (the version for the PDOK/NGR structures) will be used
}
\examples{
CSW_set_version()
CSW_set_version('2.0.2')
}
\seealso{
\code{\link{CSW_get_version}}
}
